Refactor internal/controllers into sub packages
