+++ internal/controllers/addon/phase_ensure_operator_group_test.go
@@ -1,14 +1,15 @@
-package controllers
+package addon
 
 import (
 	"context"
 	"testing"
 
+	operatorsv1 "github.com/operator-framework/api/pkg/operators/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/api/errors"
-
+	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -17,18 +18,8 @@ import (
 	"github.com/openshift/addon-operator/internal/testutil"
 )
 
-func TestEnsureAddonInstance(t *testing.T) {
-	t.Run("ensures AddonInstance", func(t *testing.T) {
-		addonInstance := &addonsv1alpha1.AddonInstance{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      addonsv1alpha1.DefaultAddonInstanceName,
-				Namespace: "addon-system",
-			},
-			Spec: addonsv1alpha1.AddonInstanceSpec{
-				HeartbeatUpdatePeriod: defaultAddonInstanceHeartbeatUpdatePeriod,
-			},
-		}
-
+func TestEnsureOperatorGroup(t *testing.T) {
+	t.Run("ensures OperatorGroup", func(t *testing.T) {
 		addonOwnNamespace := &addonsv1alpha1.Addon{
 			ObjectMeta: metav1.ObjectMeta{
 				Name: "addon-1",
@@ -88,7 +79,7 @@ func TestEnsureAddonInstance(t *testing.T) {
 				c := testutil.NewClient()
 				r := AddonReconciler{
 					Client: c,
-					Scheme: newTestSchemeWithAddonsv1alpha1(),
+					Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 				}
 				addon := test.addon
 
@@ -104,70 +95,45 @@ func TestEnsureAddonInstance(t *testing.T) {
 						mock.Anything,
 					).
 					Return(errors.NewNotFound(schema.GroupResource{}, ""))
-				var createdAddonInstance *addonsv1alpha1.AddonInstance
+				var createdOpeatorGroup *operatorsv1.OperatorGroup
 				c.
 					On(
 						"Create",
 						mock.Anything,
-						mock.IsType(&addonsv1alpha1.AddonInstance{}),
+						mock.IsType(&operatorsv1.OperatorGroup{}),
 						mock.Anything,
 					).
 					Run(func(args mock.Arguments) {
-						createdAddonInstance = args.Get(1).(*addonsv1alpha1.AddonInstance)
-					}).
-					Return(nil)
-
-				c.
-					On(
-						"Get",
-						mock.Anything,
-						client.ObjectKeyFromObject(addonInstance),
-						mock.IsType(&addonsv1alpha1.AddonInstance{}),
-					).
-					Return(nil)
-
-				c.
-					On(
-						"Update",
-						mock.Anything,
-						mock.IsType(&addonsv1alpha1.AddonInstance{}),
-						mock.Anything,
-					).
-					Run(func(args mock.Arguments) {
-						createdAddonInstance = args.Get(1).(*addonsv1alpha1.AddonInstance)
+						createdOpeatorGroup = args.Get(1).(*operatorsv1.OperatorGroup)
 					}).
 					Return(nil)
 
 				// Test
 				ctx := context.Background()
-				err := r.ensureAddonInstance(ctx, log, addon)
+				stop, err := r.ensureOperatorGroup(ctx, log, addon)
 				require.NoError(t, err)
+				assert.False(t, stop)
 
-				assert.Equal(t, addonsv1alpha1.DefaultAddonInstanceName, createdAddonInstance.Name)
-				assert.Equal(t, test.targetNamespace, createdAddonInstance.Namespace)
-				assert.Equal(t, defaultAddonInstanceHeartbeatUpdatePeriod, createdAddonInstance.Spec.HeartbeatUpdatePeriod)
+				if c.AssertCalled(
+					t, "Create",
+					mock.Anything,
+					mock.IsType(&operatorsv1.OperatorGroup{}),
+					mock.Anything,
+				) {
+					assert.Equal(t, addon.Name, createdOpeatorGroup.Name)
+					assert.Equal(t, test.targetNamespace, createdOpeatorGroup.Namespace)
+
+					assert.Equal(t, test.expectedTargetNamespaces, createdOpeatorGroup.Spec.TargetNamespaces)
+				}
 			})
 		}
 	})
 
-	t.Run("gracefully handles invalid configuration", func(t *testing.T) {
+	t.Run("guards against invalid configuration", func(t *testing.T) {
 		tests := []struct {
 			name  string
 			addon *addonsv1alpha1.Addon
 		}{
-			{
-				name: "install.type is unsupported",
-				addon: &addonsv1alpha1.Addon{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "addon-1",
-					},
-					Spec: addonsv1alpha1.AddonSpec{
-						Install: addonsv1alpha1.AddonInstallSpec{
-							Type: addonsv1alpha1.AddonInstallType("random"),
-						},
-					},
-				},
-			},
 			{
 				name: "ownNamespace is nil",
 				addon: &addonsv1alpha1.Addon{
@@ -230,7 +196,7 @@ func TestEnsureAddonInstance(t *testing.T) {
 				c := testutil.NewClient()
 				r := AddonReconciler{
 					Client: c,
-					Scheme: newTestSchemeWithAddonsv1alpha1(),
+					Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 				}
 
 				// Mock Setup
@@ -245,46 +211,85 @@ func TestEnsureAddonInstance(t *testing.T) {
 
 				// Test
 				ctx := context.Background()
-				err := r.ensureAddonInstance(ctx, log, test.addon)
-				require.EqualError(t, err, "failed to create addonInstance due to misconfigured install.spec.type")
+				stop, err := r.ensureOperatorGroup(ctx, log, test.addon)
+				require.NoError(t, err)
+				assert.True(t, stop)
+
+				c.StatusMock.AssertCalled(
+					t, "Update", mock.Anything, test.addon, mock.Anything)
+
+				availableCond := meta.FindStatusCondition(test.addon.Status.Conditions, addonsv1alpha1.Available)
+				if assert.NotNil(t, availableCond) {
+					assert.Equal(t, metav1.ConditionFalse, availableCond.Status)
+					assert.Equal(t, "ConfigurationError", availableCond.Reason)
+				}
 			})
 		}
 	})
-}
 
-func TestReconcileAddonInstance(t *testing.T) {
-	addonInstance := &addonsv1alpha1.AddonInstance{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      addonsv1alpha1.DefaultAddonInstanceName,
-			Namespace: "test",
-		},
-		Spec: addonsv1alpha1.AddonInstanceSpec{
-			HeartbeatUpdatePeriod: defaultAddonInstanceHeartbeatUpdatePeriod,
-		},
-	}
+	t.Run("unsupported install type", func(t *testing.T) {
+		addonUnsupported := &addonsv1alpha1.Addon{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "addon-1",
+			},
+			Spec: addonsv1alpha1.AddonSpec{
+				Install: addonsv1alpha1.AddonInstallSpec{
+					Type: addonsv1alpha1.AddonInstallType("something something"),
+				},
+			},
+		}
 
-	t.Run("no addoninstance", func(t *testing.T) {
+		log := testutil.NewLogger(t)
 		c := testutil.NewClient()
 		r := AddonReconciler{
 			Client: c,
-			Scheme: newTestSchemeWithAddonsv1alpha1(),
+			Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
+		}
+
+		// Test
+		ctx := context.Background()
+		stop, err := r.ensureOperatorGroup(ctx, log, addonUnsupported.DeepCopy())
+		require.NoError(t, err)
+		assert.True(t, stop)
+
+		// indirect sanity check
+		// nothing was called on the client and the method signals to stop
+	})
+}
+
+func TestReconcileOperatorGroup(t *testing.T) {
+	operatorGroup := &operatorsv1.OperatorGroup{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "testing",
+			Namespace: "testing-ns",
+		},
+		Spec: operatorsv1.OperatorGroupSpec{
+			TargetNamespaces: []string{"testing-ns"},
+		},
+	}
+
+	t.Run("no-op", func(t *testing.T) {
+		c := testutil.NewClient()
+		r := AddonReconciler{
+			Client: c,
+			Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 		}
 
 		c.
 			On(
 				"Get",
 				mock.Anything,
-				client.ObjectKeyFromObject(addonInstance),
-				mock.IsType(&addonsv1alpha1.AddonInstance{}),
+				client.ObjectKeyFromObject(operatorGroup),
+				mock.IsType(&operatorsv1.OperatorGroup{}),
 			).
 			Run(func(args mock.Arguments) {
-				fetchedAddonInstance := args.Get(2).(*addonsv1alpha1.AddonInstance)
-				addonInstance.DeepCopyInto(fetchedAddonInstance)
+				og := args.Get(2).(*operatorsv1.OperatorGroup)
+				operatorGroup.DeepCopyInto(og)
 			}).
 			Return(nil)
 
 		ctx := context.Background()
-		err := r.reconcileAddonInstance(ctx, addonInstance.DeepCopy())
+		err := r.reconcileOperatorGroup(ctx, operatorGroup.DeepCopy())
 		require.NoError(t, err)
 	})
 
@@ -292,15 +297,15 @@ func TestReconcileAddonInstance(t *testing.T) {
 		c := testutil.NewClient()
 		r := AddonReconciler{
 			Client: c,
-			Scheme: newTestSchemeWithAddonsv1alpha1(),
+			Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 		}
 
 		c.
 			On(
 				"Get",
 				mock.Anything,
-				client.ObjectKeyFromObject(addonInstance),
-				mock.IsType(&addonsv1alpha1.AddonInstance{}),
+				client.ObjectKeyFromObject(operatorGroup),
+				mock.IsType(&operatorsv1.OperatorGroup{}),
 			).
 			Return(nil)
 
@@ -308,19 +313,19 @@ func TestReconcileAddonInstance(t *testing.T) {
 			On(
 				"Update",
 				mock.Anything,
-				mock.IsType(&addonsv1alpha1.AddonInstance{}),
+				mock.IsType(&operatorsv1.OperatorGroup{}),
 				mock.Anything,
 			).
 			Return(nil)
 
 		ctx := context.Background()
-		err := r.reconcileAddonInstance(ctx, addonInstance.DeepCopy())
+		err := r.reconcileOperatorGroup(ctx, operatorGroup.DeepCopy())
 		require.NoError(t, err)
 
 		c.AssertCalled(t,
 			"Update",
 			mock.Anything,
-			mock.IsType(&addonsv1alpha1.AddonInstance{}),
+			mock.IsType(&operatorsv1.OperatorGroup{}),
 			mock.Anything,
 		)
 	})
