+++ internal/controllers/addon/phase_ensure_wanted_namespaces_test.go
@@ -1,4 +1,4 @@
-package controllers
+package addon
 
 import (
 	"context"
@@ -12,6 +12,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	addonsv1alpha1 "github.com/openshift/addon-operator/apis/addons/v1alpha1"
+	"github.com/openshift/addon-operator/internal/controllers"
 	"github.com/openshift/addon-operator/internal/testutil"
 )
 
@@ -21,11 +22,11 @@ func TestEnsureWantedNamespaces_AddonWithoutNamespaces(t *testing.T) {
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithoutNamespace())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithoutNamespace())
 	require.NoError(t, err)
 	require.False(t, stop)
 	c.AssertExpectations(t)
@@ -35,18 +36,18 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_Collision(t *testing.T)
 	c := testutil.NewClient()
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Run(func(args mock.Arguments) {
 		arg := args.Get(2).(*corev1.Namespace)
-		newTestExistingNamespaceWithOwner().DeepCopyInto(arg)
+		testutil.NewTestExistingNamespaceWithOwner().DeepCopyInto(arg)
 	}).Return(nil)
 	c.StatusMock.On("Update", testutil.IsContext, testutil.IsAddonsv1alpha1AddonPtr, mock.Anything).Return(nil)
 
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithSingleNamespace())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithSingleNamespace())
 	require.NoError(t, err)
 	require.True(t, stop)
 	c.AssertExpectations(t)
@@ -56,7 +57,7 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_Collision(t *testing.T)
 
 func TestEnsureWantedNamespaces_AddonWithSingleNamespace_NoCollision(t *testing.T) {
 	c := testutil.NewClient()
-	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(newTestErrNotFound())
+	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(testutil.NewTestErrNotFound())
 	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Run(func(args mock.Arguments) {
 		arg := args.Get(1).(*corev1.Namespace)
 		arg.Status = corev1.NamespaceStatus{
@@ -67,11 +68,11 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_NoCollision(t *testing.
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithSingleNamespace())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithSingleNamespace())
 	require.NoError(t, err)
 	require.False(t, stop)
 	c.AssertExpectations(t)
@@ -81,7 +82,7 @@ func TestEnsureWantedNamespaces_AddonWithSingleNamespace_NoCollision(t *testing.
 
 func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_NoCollision(t *testing.T) {
 	c := testutil.NewClient()
-	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(newTestErrNotFound())
+	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(testutil.NewTestErrNotFound())
 	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Run(func(args mock.Arguments) {
 		arg := args.Get(1).(*corev1.Namespace)
 		arg.Status = corev1.NamespaceStatus{
@@ -92,15 +93,15 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_NoCollision(t *testi
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithMultipleNamespaces())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithMultipleNamespaces())
 	require.NoError(t, err)
 	require.False(t, stop)
 	// every namespace should have been created
-	namespaceCount := len(newTestAddonWithMultipleNamespaces().Spec.Namespaces)
+	namespaceCount := len(testutil.NewTestAddonWithMultipleNamespaces().Spec.Namespaces)
 	c.AssertExpectations(t)
 	c.AssertNumberOfCalls(t, "Get", namespaceCount)
 	c.AssertNumberOfCalls(t, "Create", namespaceCount)
@@ -109,7 +110,7 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_NoCollision(t *testi
 func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_SingleCollision(t *testing.T) {
 	c := testutil.NewClient()
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).
-		Return(newTestErrNotFound()).
+		Return(testutil.NewTestErrNotFound()).
 		Once()
 	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).
 		Run(func(args mock.Arguments) {
@@ -123,7 +124,7 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_SingleCollision(t *t
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).
 		Run(func(args mock.Arguments) {
 			arg := args.Get(2).(*corev1.Namespace)
-			newTestExistingNamespaceWithOwner().DeepCopyInto(arg)
+			testutil.NewTestExistingNamespaceWithOwner().DeepCopyInto(arg)
 		}).
 		Return(nil)
 	c.StatusMock.On("Update", testutil.IsContext, testutil.IsAddonsv1alpha1AddonPtr, mock.Anything).
@@ -132,15 +133,15 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_SingleCollision(t *t
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithMultipleNamespaces())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithMultipleNamespaces())
 	require.NoError(t, err)
 	require.True(t, stop)
 	c.AssertExpectations(t)
-	c.AssertNumberOfCalls(t, "Get", len(newTestAddonWithMultipleNamespaces().Spec.Namespaces))
+	c.AssertNumberOfCalls(t, "Get", len(testutil.NewTestAddonWithMultipleNamespaces().Spec.Namespaces))
 	c.StatusMock.AssertCalled(t, "Update", testutil.IsContext, testutil.IsAddonsv1alpha1AddonPtr, mock.Anything)
 
 }
@@ -149,7 +150,7 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_MultipleCollisions(t
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).
 		Run(func(args mock.Arguments) {
 			arg := args.Get(2).(*corev1.Namespace)
-			newTestExistingNamespaceWithOwner().DeepCopyInto(arg)
+			testutil.NewTestExistingNamespaceWithOwner().DeepCopyInto(arg)
 		}).
 		Return(nil)
 	c.StatusMock.On("Update", testutil.IsContext, testutil.IsAddonsv1alpha1AddonPtr, mock.Anything).
@@ -158,29 +159,29 @@ func TestEnsureWantedNamespaces_AddonWithMultipleNamespaces_MultipleCollisions(t
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
-	stop, err := r.ensureWantedNamespaces(ctx, newTestAddonWithMultipleNamespaces())
+	stop, err := r.ensureWantedNamespaces(ctx, testutil.NewTestAddonWithMultipleNamespaces())
 	require.NoError(t, err)
 	require.True(t, stop)
 	c.AssertExpectations(t)
-	c.AssertNumberOfCalls(t, "Get", len(newTestAddonWithMultipleNamespaces().Spec.Namespaces))
+	c.AssertNumberOfCalls(t, "Get", len(testutil.NewTestAddonWithMultipleNamespaces().Spec.Namespaces))
 	c.StatusMock.AssertCalled(t, "Update", testutil.IsContext, testutil.IsAddonsv1alpha1AddonPtr, mock.Anything)
 }
 
 func TestEnsureNamespace_Create(t *testing.T) {
-	addon := newTestAddonWithSingleNamespace()
+	addon := testutil.NewTestAddonWithSingleNamespace()
 
 	c := testutil.NewClient()
-	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(newTestErrNotFound())
+	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(testutil.NewTestErrNotFound())
 	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Return(nil)
 
 	r := &AddonReconciler{
 		Client: c,
 		Log:    testutil.NewLogger(t),
-		Scheme: newTestSchemeWithAddonsv1alpha1(),
+		Scheme: testutil.NewTestSchemeWithAddonsv1alpha1(),
 	}
 
 	ctx := context.Background()
@@ -192,33 +193,33 @@ func TestEnsureNamespace_Create(t *testing.T) {
 
 func TestReconcileNamespace_Create(t *testing.T) {
 	c := testutil.NewClient()
-	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(newTestErrNotFound())
-	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Return(nil, newTestNamespace())
+	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Return(testutil.NewTestErrNotFound())
+	c.On("Create", testutil.IsContext, testutil.IsCoreV1NamespacePtr, mock.Anything).Return(nil, testutil.NewTestNamespace())
 
 	ctx := context.Background()
-	reconciledNamespace, err := reconcileNamespace(ctx, c, newTestSchemeWithAddonsv1alpha1(), newTestNamespace(),
+	reconciledNamespace, err := reconcileNamespace(ctx, c, testutil.NewTestSchemeWithAddonsv1alpha1(), testutil.NewTestNamespace(),
 		addonsv1alpha1.ResourceAdoptionPrevent)
 	require.NoError(t, err)
 	assert.NotNil(t, reconciledNamespace)
-	assert.Equal(t, newTestNamespace(), reconciledNamespace)
+	assert.Equal(t, testutil.NewTestNamespace(), reconciledNamespace)
 	c.AssertExpectations(t)
 	c.AssertCalled(t, "Get", testutil.IsContext, client.ObjectKey{
 		Name: "namespace-1",
 	}, testutil.IsCoreV1NamespacePtr)
-	c.AssertCalled(t, "Create", testutil.IsContext, newTestNamespace(), mock.Anything)
+	c.AssertCalled(t, "Create", testutil.IsContext, testutil.NewTestNamespace(), mock.Anything)
 }
 
 func TestReconcileNamespace_CreateWithCollisionWithoutOwner(t *testing.T) {
 	c := testutil.NewClient()
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Run(func(args mock.Arguments) {
 		arg := args.Get(2).(*corev1.Namespace)
-		newTestExistingNamespaceWithoutOwner().DeepCopyInto(arg)
+		testutil.NewTestExistingNamespaceWithoutOwner().DeepCopyInto(arg)
 	}).Return(nil)
 
 	ctx := context.Background()
-	_, err := reconcileNamespace(ctx, c, newTestSchemeWithAddonsv1alpha1(), newTestNamespace(),
+	_, err := reconcileNamespace(ctx, c, testutil.NewTestSchemeWithAddonsv1alpha1(), testutil.NewTestNamespace(),
 		addonsv1alpha1.ResourceAdoptionPrevent)
-	require.EqualError(t, err, errNotOwnedByUs.Error())
+	require.EqualError(t, err, controllers.ErrNotOwnedByUs.Error())
 	c.AssertExpectations(t)
 	c.AssertCalled(t, "Get", testutil.IsContext, client.ObjectKey{
 		Name: "namespace-1",
@@ -229,13 +230,13 @@ func TestReconcileNamespace_CreateWithCollisionWithOtherOwner(t *testing.T) {
 	c := testutil.NewClient()
 	c.On("Get", testutil.IsContext, testutil.IsObjectKey, testutil.IsCoreV1NamespacePtr).Run(func(args mock.Arguments) {
 		arg := args.Get(2).(*corev1.Namespace)
-		newTestExistingNamespaceWithoutOwner().DeepCopyInto(arg)
+		testutil.NewTestExistingNamespaceWithoutOwner().DeepCopyInto(arg)
 	}).Return(nil)
 
 	ctx := context.Background()
-	_, err := reconcileNamespace(ctx, c, newTestSchemeWithAddonsv1alpha1(), newTestNamespace(),
+	_, err := reconcileNamespace(ctx, c, testutil.NewTestSchemeWithAddonsv1alpha1(), testutil.NewTestNamespace(),
 		addonsv1alpha1.ResourceAdoptionPrevent)
-	require.EqualError(t, err, errNotOwnedByUs.Error())
+	require.EqualError(t, err, controllers.ErrNotOwnedByUs.Error())
 	c.AssertExpectations(t)
 	c.AssertCalled(t, "Get", testutil.IsContext, client.ObjectKey{
 		Name: "namespace-1",
@@ -250,7 +251,7 @@ func TestReconcileNamespace_CreateWithClientError(t *testing.T) {
 		Return(timeoutErr)
 
 	ctx := context.Background()
-	_, err := reconcileNamespace(ctx, c, newTestSchemeWithAddonsv1alpha1(), newTestNamespace(),
+	_, err := reconcileNamespace(ctx, c, testutil.NewTestSchemeWithAddonsv1alpha1(), testutil.NewTestNamespace(),
 		addonsv1alpha1.ResourceAdoptionPrevent)
 	require.Error(t, err)
 	require.EqualError(t, err, timeoutErr.Error())
@@ -258,21 +259,4 @@ func TestReconcileNamespace_CreateWithClientError(t *testing.T) {
 	c.AssertCalled(t, "Get", testutil.IsContext, client.ObjectKey{
 		Name: "namespace-1",
 	}, testutil.IsCoreV1NamespacePtr)
-}
-
-func TestHasEqualControllerReference(t *testing.T) {
-	require.True(t, HasEqualControllerReference(
-		newTestNamespace(),
-		newTestNamespace(),
-	))
-
-	require.False(t, HasEqualControllerReference(
-		newTestNamespace(),
-		newTestExistingNamespaceWithOwner(),
-	))
-
-	require.False(t, HasEqualControllerReference(
-		newTestNamespace(),
-		newTestExistingNamespaceWithoutOwner(),
-	))
 }
