+++ internal/controllers/addonoperator/utils.go
@@ -1,4 +1,4 @@
-package controllers
+package addonoperator
 
 import (
 	"context"
