+++ internal/controllers/common_utils_test.go
@@ -3,10 +3,29 @@ package controllers
 import (
 	"testing"
 
+	"github.com/stretchr/testify/require"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	addonsv1alpha1 "github.com/openshift/addon-operator/apis/addons/v1alpha1"
+	"github.com/openshift/addon-operator/internal/testutil"
 )
+
+func TestHasEqualControllerReference(t *testing.T) {
+	require.True(t, HasEqualControllerReference(
+		testutil.NewTestNamespace(),
+		testutil.NewTestNamespace(),
+	))
+
+	require.False(t, HasEqualControllerReference(
+		testutil.NewTestNamespace(),
+		testutil.NewTestExistingNamespaceWithOwner(),
+	))
+
+	require.False(t, HasEqualControllerReference(
+		testutil.NewTestNamespace(),
+		testutil.NewTestExistingNamespaceWithoutOwner(),
+	))
+}
 
 func TestAddCommonLabels(t *testing.T) {
 	addon := &addonsv1alpha1.Addon{
@@ -17,7 +36,7 @@ func TestAddCommonLabels(t *testing.T) {
 
 	labels := make(map[string]string)
 
-	addCommonLabels(labels, addon)
+	AddCommonLabels(labels, addon)
 
 	if labels[commonInstanceLabel] != addon.Name {
 		t.Error("commonInstanceLabel was not set to addon name")
@@ -34,7 +53,7 @@ func TestCommonLabelsAsLabelSelector(t *testing.T) {
 			Name: "test",
 		},
 	}
-	selector := commonLabelsAsLabelSelector(addonWithCorrectName)
+	selector := CommonLabelsAsLabelSelector(addonWithCorrectName)
 
 	if selector.Empty() {
 		t.Fatal("selector is empty but should filter on common labels")
