+++ internal/controllers/errors.go
@@ -5,5 +5,5 @@ import "errors"
 var (
 	// This error is returned when a reconciled child object already
 	// exists and is not owned by the current controller/addon
-	errNotOwnedByUs = errors.New("object is not owned by us")
+	ErrNotOwnedByUs = errors.New("object is not owned by us")
 )
